+++ lib/src/meteor/subscribed_collection.dart
@@ -33,12 +33,12 @@ class SubscribedCollection {
 
   /// Returns specific objects from a subscribed collection using a set of [selectors].
   Map<String, Map<String, dynamic>> find(Map<String, dynamic> selectors) {
-    Map<String, Map<String, dynamic>> filteredCollection =
+    var filteredCollection =
         <String, Map<String, dynamic>>{};
     print('Finding docs');
     print(selectors.keys);
     _collection.findAll().forEach((key, document) {
-      bool shouldAdd = true;
+      var shouldAdd = true;
       selectors.forEach((selector, value) {
         print('Key: $selector');
         print('Value: ${document[selector]}');