+++ lib/src/utils/utils.dart
@@ -18,7 +18,7 @@ class Utils {
   }
 
   static String _decodeBase64(String str) {
-    String output = str.replaceAll('-', '+').replaceAll('_', '/');
+    var output = str.replaceAll('-', '+').replaceAll('_', '/');
 
     switch (output.length % 4) {
       case 0: